+++ cli/src/template.rs
@@ -79,6 +79,24 @@ module.exports = async function (provider) {
 }
 "#
 }
+
+pub fn ts_deploy_script() -> &'static str {
+    r#"
+// Migrations are an early feature. Currently, they're nothing more than this
+// single deploy script that's invoked from the CLI, injecting a provider
+// configured from the workspace's Anchor.toml.
+
+const anchor = require("@project-serum/anchor");
+
+module.exports = async function (provider) {
+  // Configure client to use the provider.
+  anchor.setProvider(provider);
+
+  // Add your deploy script here.
+}
+"#
+}
+
 pub fn xargo_toml() -> &'static str {
     r#"[target.bpfel-unknown-unknown.dependencies.std]
 features = []"#